it works but barely

+++ libs/sdw/RayTriangleIntersection.h
@@ -8,7 +8,7 @@ struct RayTriangleIntersection {
 	glm::vec3 intersectionPoint;
 	float distanceFromCamera;
 	ModelTriangle intersectedTriangle;
-	size_t triangleIndex;
+	long triangleIndex;
 
 	RayTriangleIntersection();
 	RayTriangleIntersection(const glm::vec3 &point, float distance, const ModelTriangle &triangle, size_t index);
